+++ CHANGELOG.rst
@@ -1,7 +1,16 @@
 Changelog
 =========
 
-v34.9.3 (unreleased)
+v34.9.4 (unreleased)
+--------------------
+
+- Improve Project list page navigation.
+  A top previous/next page navigation was added in the header for consistency with other
+  list views.
+  Any paginated view can now be navigated using the left/right keyboard keys.
+  https://github.com/aboutcode-org/scancode.io/issues/1200
+
+v34.9.3 (2024-12-31)
 --------------------
 
 - Refine the available settings for RQ_QUEUES: